% file MASS/man/abbey.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{abbey}
\alias{abbey}
\title{
  Determinations of Nickel Content
}
\description{
  A numeric vector of 31 determinations of nickel content (ppm) in
  a Canadian syenite rock.
}
\usage{
abbey
}
\source{
  S. Abbey (1988) \emph{Geostandards Newsletter} \bold{12}, 241.
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
