% file MASS/man/animals.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{Animals}
\alias{Animals}
\title{
  Brain and Body Weights for 28 Species
}
\description{
  Average brain and body weights for 28 species of land animals.
}
\usage{
Animals
}
\format{
  \describe{
    \item{\code{body}}{
      body weight in kg.
    }
    \item{\code{brain}}{
      brain weight in g.
    }
  }
}
\note{
  The name \code{Animals} avoids conflicts with a system dataset
  \code{animals} in S-PLUS 4.5 and later.
}
\source{
  P. J. Rousseeuw  and A. M. Leroy (1987)
  \emph{Robust Regression and Outlier Detection.}
  Wiley, p. 57.
}
\references{
  Venables, W. N. and Ripley, B. D. (1999)
  \emph{Modern Applied Statistics with S-PLUS.} Third
  Edition. Springer.
}
\keyword{datasets}
