% file MASS/chem.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{chem}
\alias{chem}
\title{
Copper in Wholemeal Flour
}
\description{
A numeric vector of 24 determinations of copper in wholemeal
flour, in parts per million.
}
\source{
Analytical Methods Committee (1989) Robust statistics -- how not to
reject outliers.
\emph{The Analyst}
\bold{114}, 1693-1702, 1989
}
\keyword{datasets}
