% file MASS/man/confint.Rd
% copyright (C) 1999-2023 W. N. Venables and B. D. Ripley
%
\name{confint-MASS}
\alias{confint.glm}
\alias{confint.nls}
\alias{confint.profile.glm}
\alias{confint.profile.nls}
\title{
Confidence Intervals for Model Parameters
}
\description{
  Computes confidence intervals for one or more parameters in a fitted
  model.  Package \pkg{MASS} added methods for \code{glm} and \code{nls}
  fits.  As fron \R 4.4.0 these have been migrated to package
  \pkg{stats}.

  It also adds a method for \code{\link{polr}} fits.
}

\keyword{models}
