% file MASS/man/DDT.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{DDT}
\alias{DDT}
\title{
DDT in Kale
}
\description{
  A numeric vector of 15 measurements by different laboratories of the
  pesticide DDT in kale, in ppm (parts per million) using the multiple
  pesticide residue measurement.
}
\usage{
DDT
}
\source{
  C. E. Finsterwalder (1976)
  Collaborative study of an extension of the Mills \emph{et al}
  method for the determination of pesticide residues in food.
  \emph{J. Off. Anal. Chem.} \bold{59}, 169--171

  R. G. Staudte and S. J. Sheather (1990)
  \emph{Robust Estimation and Testing.}
  Wiley
}
\keyword{datasets}
