% file MASS/man/deaths.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{deaths}
\alias{deaths}
\title{
  Monthly Deaths from Lung Diseases in the UK
}
\description{
  A time series giving the monthly deaths from bronchitis,
  emphysema and asthma in the UK, 1974-1979, both sexes (\code{deaths}),
}
\usage{
deaths
}
\source{
  P. J. Diggle (1990)
  \emph{Time Series: A Biostatistical Introduction.}
  Oxford, table A.3
}
\seealso{
  This the same as dataset \code{\link{ldeaths}} in \R's \pkg{datasets} package.
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
