% file MASS/man/drivers.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{drivers}
\alias{drivers}
\title{
  Deaths of Car Drivers in Great Britain 1969-84
}
\description{
  A regular time series giving the monthly totals of car drivers in
  Great Britain killed or seriously injured Jan 1969 to Dec
  1984.  Compulsory wearing of seat belts was introduced on 31 Jan 1983
}
\usage{
drivers
}
\source{
  Harvey, A.C. (1989)
  \emph{Forecasting, Structural Time Series Models and the Kalman Filter.}
  Cambridge University Press, pp. 519--523.
}
\references{
  Venables, W. N. and Ripley, B. D. (1999)
  \emph{Modern Applied Statistics with S-PLUS.} Third
  Edition. Springer.
}
\keyword{datasets}
