% file MASS/man/forbes.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{forbes}
\alias{forbes}
\title{
  Forbes' Data on Boiling Points in the Alps
}
\description{
  A data frame with 17 observations on boiling point
  of water and barometric pressure in inches of mercury.
}
\usage{
forbes
}
\format{
  \describe{
    \item{\code{bp}}{
      boiling point (degrees Farenheit).
    }
    \item{\code{pres}}{
      barometric pressure in inches of mercury.
    }
  }
}
\source{
  A. C. Atkinson (1985) \emph{Plots, Transformations and Regression.} Oxford.

  S. Weisberg (1980) \emph{Applied Linear Regression.} Wiley.
}
\keyword{datasets}
