% file MASS/man/GAGurine.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{GAGurine}
\alias{GAGurine}
\title{
  Level of GAG in Urine of Children
}
\description{
  Data were collected on the concentration of a chemical GAG in the
  urine of 314 children aged from zero to seventeen years.  The aim of
  the study was to produce a chart to help a paediatrican to assess if a
  child's GAG concentration is \sQuote{normal}.
}
\usage{
GAGurine
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{\code{Age}}{
      age of child in years.
    }
    \item{\code{GAG}}{
      concentration of GAG (the units have been lost).
    }
  }
}
\source{
  Mrs Susan Prosser, Paediatrics Department, University of Oxford,
  via Department of Statistics Consulting Service.
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
