% file MASS/man/mammals.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{mammals}
\alias{mammals}
\title{
  Brain and Body Weights for 62 Species of Land Mammals
}
\description{
  A data frame with average brain and body weights for 62 species
  of land mammals.
}
\usage{
mammals
}
\format{
  \describe{
    \item{\code{body}}{
      body weight in kg.
    }
    \item{\code{brain}}{
      brain weight in g.
    }
    \item{\code{name}}{
      Common name of species.
      (Rock hyrax-a = \emph{Heterohyrax brucci},
      Rock hyrax-b = \emph{Procavia habessinic.}.)
    }
  }
}
\source{
  Weisberg, S. (1985)
  \emph{Applied Linear Regression.} 2nd edition. Wiley, pp. 144--5.

  Selected from:
  Allison, T. and Cicchetti, D. V. (1976)
  Sleep in mammals: ecological and constitutional correlates.
  \emph{Science} \bold{194}, 732--734.
}
\references{
  Venables, W. N. and Ripley, B. D. (1999)
  \emph{Modern Applied Statistics with S-PLUS.} Third
  Edition. Springer.
}
\keyword{datasets}
