% file MASS/phones.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{phones}
\alias{phones}
\title{
Belgium Phone Calls 1950-1973
}
\description{
A list object with the annual numbers of telephone calls, in
Belgium.  The components are: \code{year}: year - 1900 and \code{calls}:
telephone calls in millions.
}
\source{
P. J. Rousseeuw and A. M. Leroy (1987)
\emph{Robust Regression & Outlier Detection.}
Wiley.
}
\keyword{datasets}
