% file MASS/man/profile.glm.Rd
% copyright (C) 1999-2023 W. N. Venables and B. D. Ripley

\name{profile.glm}
\alias{profile.glm}
\title{Method for Profiling glm Objects}
\description{
  Investigates the profile log-likelihood function for a fitted model of
  class \code{"glm"}.

  As from \R 4.4.0 was migrated to package \pkg{stats} with additional
  functionality.
}

\keyword{regression}
\keyword{models}
