% file MASS/man/road.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{road}
\alias{road}
\title{
  Road Accident Deaths in US States
}
\description{
  A data frame with the annual deaths in road accidents for half
  the US states.
}
\usage{
road
}
\format{
  Columns are:
  \describe{
    \item{\code{state}}{
      name.
    }
    \item{\code{deaths}}{
      number of deaths.
    }
    \item{\code{drivers}}{
      number of drivers (in 10,000s).
    }
    \item{\code{popden}}{
      population density in people per square mile.
    }
    \item{\code{rural}}{
      length of rural roads, in 1000s of miles.
    }
    \item{\code{temp}}{
      average daily maximum temperature in January.
    }
    \item{\code{fuel}}{
      fuel consumption in 10,000,000 US gallons per year.
    }
  }
}
\source{
  Imperial College, London M.Sc. exercise
}
\keyword{datasets}
