% file MASS/man/rotifer.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{rotifer}
\alias{rotifer}
\title{
  Numbers of Rotifers by Fluid Density
}
\description{
  The data give the numbers of rotifers falling out of suspension for
  different fluid densities. There are two species, \code{pm}
  \emph{Polyartha major} and \code{kc}, \emph{Keratella cochlearis} and
  for each species the number falling out and the total number are
  given.
}
\usage{
rotifer
}
\format{
  \describe{
    \item{\code{density}}{
      specific density of fluid.
    }
    \item{\code{pm.y}}{
      number falling out for \emph{P. major}.
    }
    \item{\code{pm.total}}{
      total number of \emph{P. major}.
    }
    \item{\code{kc.y}}{
      number falling out for \emph{K. cochlearis}.
    }
    \item{\code{kc.tot}}{
      total number of \emph{K. cochlearis}.
    }
  }
}
\source{
  D. Collett (1991) \emph{Modelling Binary Data.} Chapman & Hall. p. 217
}
\keyword{datasets}
