% file MASS/man/Rubber.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{Rubber}
\alias{Rubber}
\title{
  Accelerated Testing of Tyre Rubber
}
\description{
  Data frame from accelerated testing of tyre rubber.
}
\usage{
Rubber
}
\format{
  \describe{
    \item{\code{loss}}{
      the abrasion loss in gm/hr.
    }
    \item{\code{hard}}{
      the hardness in Shore units.
    }
    \item{\code{tens}}{
      tensile strength in kg/sq m.
    }
  }
}
\source{
  O.L. Davies (1947)
  \emph{Statistical Methods in Research and Production.}
  Oliver and Boyd, Table 6.1 p. 119.

  O.L. Davies and P.L. Goldsmith (1972)
  \emph{Statistical Methods in Research and Production.}
  4th edition, Longmans, Table 8.1 p. 239.
}
\references{
  Venables, W. N. and Ripley, B. D. (1999)
  \emph{Modern Applied Statistics with S-PLUS.} Third
  Edition. Springer.
}
\keyword{datasets}
