% file MASS/man/shrimp.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{shrimp}
\alias{shrimp}
\title{
  Percentage of Shrimp in Shrimp Cocktail
}
\description{
  A numeric vector with 18 determinations by different laboratories
  of the amount (percentage of the declared total weight) of shrimp
  in shrimp cocktail.
}
\usage{
shrimp
}
\source{
  F. J. King and J. J. Ryan (1976)
  Collaborative study of the determination of the amount of shrimp in
  shrimp cocktail. \emph{J. Off. Anal. Chem.} \bold{59}, 644--649.

  R. G. Staudte and S. J. Sheather (1990)
  \emph{Robust Estimation and Testing.} Wiley.
}
\keyword{datasets}
