\name{heart}
\alias{heart}
\title{
  Veteran's Administration lung cancer trial.
}
\description{
  Veteran's Administration lung cancer trial from Kalbfleisch & Prentice.
  A data frame.
}
\details{
The components are:
}
\arguments{
\item{stime}{
  survival or follow-up time in days.
}
\item{status}{
  dead or censored.
}
\item{treat}{
  standard or test
}
\item{age}{
patient's age in years
}
\item{Karn}{
  Karnofsky score of patient's performance on a scale of 0 to 100.
}
\item{diag.time}{
  times since diagnosis in months at entry to trial.
}
\item{cell}{
  one of four cell types.
}
\item{prior}{
  prior therapy?
}}
\source{
Kalbfleisch, J.D. and Prentice R.L. (1980)
\emph{The Statistical Analysis of Failure Time Data.} 
}
\keyword{datasets}
