% file MASS/animals.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{Animals}
\alias{Animals}
\alias{animals}
\title{
Brain and Body Weights for 28 Species
}
\description{
Average brain and body weights for 28 species of land animals.
}
\format{
}
\arguments{
\item{body}{
body weight in kg
}
\item{brain}{
brain weight in g
}}
\section{NOTE}{
The name \code{Animals} avoids conflicts with a system dataset in S-PLUS
4.5 and later.
}
\source{
P. J. Rousseeuw,  and A. M. Leroy (1987)
\emph{Robust Regression and Outlier Detection.}
Wiley, p. 57.
}
\keyword{datasets}
