% file MASS/Boston.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{Boston}
\alias{Boston}
\title{
Housing Values in Suburbs of Boston
}
\description{
The \code{Boston} data frame has 506 rows and 14 columns.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{crim}{
per capita crime rate by town
}
\item{zn}{
proportion of residential land zoned for lots over 25,000 sq.ft.
}
\item{indus}{
proportion of non-retail business acres per town
}
\item{chas}{
Charles River dummy variable (= 1 if tract bounds river; 0 otherwise)
}
\item{nox}{
nitrogen oxides concentration (parts per 10 million)
}
\item{rm}{
average number of rooms per dwelling
}
\item{age}{
proportion of owner-occupied units built prior to 1940
}
\item{dis}{
weighted mean of distances to five Boston employment centres
}
\item{rad}{
index of accessibility to radial highways
}
\item{tax}{
full-value property-tax rate per $10,000
}
\item{ptratio}{
pupil-teacher ratio by town
}
\item{black}{
1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town
}
\item{lstat}{
lower status of the population (percent)
}
\item{medv}{
median value of owner-occupied homes in $1000
}}
\source{
Harrison, D. and Rubinfeld, D.L. (1978)
Hedonic prices and the demand for clean air.
\emph{J. Environ. Economics and Management}
\bold{5}, 81-102.


Belsley D.A., Kuh, E.  and Welsch, R.E. (1980)
\emph{Regression Diagnostics. Identifying Influential Data and Sources of Collinearity.}
New York: Wiley.
}
\keyword{datasets}
