% file MASS/man/chem.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{chem}
\alias{chem}
\title{
  Copper in Wholemeal Flour
}
\description{
  A numeric vector of 24 determinations of copper in wholemeal
  flour, in parts per million.
}
\usage{
chem
}
\source{
  Analytical Methods Committee (1989) Robust statistics -- how not to
  reject outliers. \emph{The Analyst} \bold{114}, 1693--1702.
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
