% file MASS/fgl.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{fgl}
\alias{fgl}
\title{
Measurements of Forensic Glass Fragments
}
\description{
The \code{fgl} data frame has 214 rows and 10 columns.
It was collected by B. German on fragments of glass
collected in forensic work.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{RI}{
refractive index; more precisely the refractive index is 1.518xxxx.


The remaining 8 measurements are percentages by weight of oxides.
}
\item{Na}{
sodium
}
\item{Mg}{
manganese
}
\item{Al}{
aluminium
}
\item{Si}{
silicon
}
\item{K}{
potassium
}
\item{Ca}{
calcium
}
\item{Ba}{
barium
}
\item{Fe}{
iron
}
\item{type}{
The fragments were originally classed into seven types, one of which
was absent in this dataset.  The categories which occur are 
window float glass (\code{WinF}: 70),
window non-float glass (\code{WinNF}: 76),
vehicle window glass (\code{Veh}: 17), 
containers (\code{Con}: 13), 
tableware (\code{Tabl}: 9) and
vehicle headlamps (\code{Head}: 29).
}}
\keyword{datasets}
