% file MASS/forbes.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{forbes}
\alias{forbes}
\title{
Forbes' Data on Boiling Points in the Alps
}
\description{
A data frame with 17 observations on boiling point (degrees F)
and barometric pressure in inches of mercury.
}
\format{
}
\arguments{
\item{bp}{
boiling point (degrees F)
}
\item{pres}{
barometric pressure in inches of mercury
}}
\source{
A. C. Atkinson (1985)
\emph{Plots, Transformations and Regression.}
Oxford.


S. Weisberg (1980)
\emph{Applied Linear Regression.}
Wiley.
}
\keyword{datasets}
