% file MASS/GAGurine.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{GAGurine}
\alias{GAGurine}
\title{
Level of GAG in Urine of Children
}
\description{
Data were collected on the concentration of a chemical GAG in the
urine of 314 children aged from zero to seventeen years.  The aim of
the study was to produce a chart to help a paediatrican to assess if a
child's GAG concentration is "normal".
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{Age}{
age of child in years
}
\item{GAG}{
concentration of GAG (the units have been lost)
}}
\source{
Mrs Susan Prosser, Paediatrics Department, University of Oxford,
via Department of Statistics Consulting Service.
}
\keyword{datasets}
