% file MASS/mcycle.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{mcycle}
\alias{mcycle}
\title{
Data from a Simulated Motorcycle Accident
}
\description{
A data frame giving a series of measurements of head acceleration
in a simulated motorcycle  
accident, used to test crash helmets.
}
\format{
}
\arguments{
\item{times}{
in milliseconds after impact
}
\item{accel}{
in g
}}
\source{
Silverman, B. W. (1985) Some aspects of the spline smoothing approach to
non-parametric curve fitting.
\emph{Journal of the Royal Statistical Society series B,}
\bold{47}, 1-52.
}
\keyword{datasets}
