% file MASS/man/newcomb.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{newcomb}
\alias{newcomb}
\title{
  Newcomb's Measurements of the Passage Time of Light
}
\description{
  A numeric vector giving the \sQuote{Third Series} of measurements of the
  passage time of light recorded by Newcomb in 1882.  The given values
  divided by 1000 plus 24 give the time in millionths of a second for
  light to traverse a known distance. The \sQuote{true} value is now
  considered to be 33.02.
}
\usage{
newcomb
}
\source{
  S. M. Stigler (1973)
  Simon Newcomb, Percy Daniell, and the history of robust estimation
  1885--1920.
  \emph{Journal of the American Statistical Association} \bold{68}, 872--879.

  R. G. Staudte and S. J. Sheather (1990)
  \emph{Robust Estimation and Testing.} Wiley.
}
\keyword{datasets}
