% file MASS/man/phones.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{Belgian-phones}
\alias{phones}
\title{
  Belgium Phone Calls 1950-1973
}
\usage{
phones
}
\description{
  A list object with the annual numbers of telephone calls, in
  Belgium.  The components are:
  \describe{
    \item{\code{year}}{
      last two digits of the year.
    }
    \item{\code{calls}}{
      number of telephone calls made (in millions of calls).
    }
  }
}
\source{
  P. J. Rousseeuw and A. M. Leroy (1987)
  \emph{Robust Regression & Outlier Detection.} Wiley.
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
