% file MASS/road.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{road}
\alias{road}
\title{
Road Accident Deaths in US States
}
\description{
A data frame with the annual deaths in road accidents for half
the US states. Components are:
}
\format{
}
\arguments{
\item{state}{
name
}
\item{deaths}{
number of deaths
}
\item{drivers}{
number of drivers (in 10,000's)
}
\item{popden}{
population density in people per square mile
}
\item{rural}{
length of rural roads, in 1000's of miles
}
\item{temp}{
average daily maximum temperature in January
}
\item{fuel}{
fuel consumption in 10,000,000 US gallons per year
}}
\source{
Imperial College, London M.Sc. exercise
}
\keyword{datasets}
