% file MASS/Rubber.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{Rubber}
\alias{Rubber}
\title{
Accelerated Testing of Tyre Rubber
}
\description{
Data frame from accelerated testing of tyre rubber. 
}
\usage{
data(Rubber)
}
\format{
\describe{
\item{\code{loss}}{
the abrasion loss in gm/hr.
}
\item{\code{hard}}{
the hardness in Shore units.
}
\item{\code{tens}}{
tensile strength in kg/sq m.
}}}
\source{
O.L. Davies (1947)
\emph{Statistical Methods in Research and Production.}
Oliver and Boyd, Table 6.1 p. 119.


O.L. Davies and P.L. Goldsmith (1972)
\emph{Statistical Methods in Research and Production.}
4th edition, Longmans, Table 8.1 p. 239.
}
\keyword{datasets}
