% file MASS/man/shoes.Rd
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{shoes}
\alias{shoes}
\title{
  Shoe wear data of Box, Hunter and Hunter
}
\description{
  A list of two vectors, giving the wear of shoes of materials A and B
  for one foot each of ten boys.
}
\usage{
shoes
}
\source{
  G. E. P. Box, W. G. Hunter and J. S. Hunter (1978)
  \emph{Statistics for Experimenters.} Wiley, p. 100
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
