\name{SP500}
\alias{SP500}
\title{
  Returns of the Standard and Poors 500
}
\description{
  Returns of the Standard and Poors 500 Index in the 1990's
}
\usage{
SP500
}
\format{
  A vector of returns of the Standard and Poors 500 index for all
  the trading days in 1990, 1991, \dots, 1999.
}
\references{
  Venables, W. N. and Ripley, B. D. (2002)
  \emph{Modern Applied Statistics with S.} Fourth edition.  Springer.
}
\keyword{datasets}
