% file MASS/steam.d
% copyright (C) 1994-9 W. N. Venables and B. D. Ripley
%
\name{steam}
\alias{steam}
\title{
The Saturated Steam Pressure Data
}
\description{
Temperature and pressure in a saturated steam driven experimental
device.
}
\usage{
data(steam)
}
\format{
The data frame contains the following components:
\describe{
\item{\code{Temp}}{
Temperature, in degrees Celsius
}
\item{\code{Press}}{
Pressure, in Pascals
}}}
\source{
N.R. Draper  and H. Smith (1981)
\emph{Applied Regression Analysis.}
Wiley, pp. 518--9.
}
\references{
  Venables, W. N. and Ripley, B. D. (1999)
  \emph{Modern Applied Statistics with S-PLUS.} Third
  Edition. Springer.
}
\keyword{datasets}
